requirements.txt
strands-agents==0.1.2
strands-agents-builder==0.1.1 
strands-agents-tools==0.1.1